requirements.txt
pandas
numpy
scipy
scikit-learn
matplotlib
llama-cpp-python
openai
